+++ src/pybamm/models/full_battery_models/lithium_ion/__init__.py
@@ -19,6 +19,7 @@ from .dfn import DFN
 from .newman_tobias import NewmanTobias
 from .basic_dfn import BasicDFN
 from .basic_spm import BasicSPM
+from .basic_spm_with_3d_thermal import Basic3DThermalSPM
 from .basic_dfn_half_cell import BasicDFNHalfCell
 from .basic_dfn_composite import BasicDFNComposite
 from .Yang2017 import Yang2017
@@ -29,4 +30,4 @@ from .basic_splitOCVR import SplitOCVR
 __all__ = ['Yang2017', 'base_lithium_ion_model', 'basic_dfn',
            'basic_dfn_composite', 'basic_dfn_half_cell', 'basic_spm', 'dfn',
            'electrode_soh', 'electrode_soh_half_cell', 'mpm', 'msmr',
-           'newman_tobias', 'spm', 'spme', 'basic_splitOCVR']
+           'newman_tobias', 'spm', 'spme', 'basic_splitOCVR', 'basic_spm_with_3d_thermal']